Move each graph type into its own file and update test cases
